refactor: move `getSshrvCommand` to common
